refactor: useParams 구현
