+++ src/pages/PostsManagerPage.tsx
@@ -11,28 +11,21 @@ import { CommentUpdateDialog } from "../components/CommentUpdateDialog"
 import { PostDetailDialog } from "../components/PostDetailDialog"
 import { UserModal } from "../components/UserModal"
 import { FilterableSearch } from "../components/FilterableSearch"
-import { getPosts, getPostsByTag } from "../api/post"
 import { getTags } from "../api/tag"
 import { usePosts } from "../hooks/usePosts"
+import { useParams } from "../hooks/useParams"
 
 const PostsManager = () => {
-  const navigate = useNavigate()
   const location = useLocation()
-  const queryParams = new URLSearchParams(location.search)
 
   // 상태 관리
+  const { limit, selectedTag, skip, sortBy, sortOrder } = useParams()
 
-  const [skip, setSkip] = useState(parseInt(queryParams.get("skip") || "0"))
-  const [limit, setLimit] = useState(parseInt(queryParams.get("limit") || "10"))
-  const [searchQuery, setSearchQuery] = useState(queryParams.get("search") || "")
   const [selectedPost, setSelectedPost] = useState(null)
-  const [sortBy, setSortBy] = useState(queryParams.get("sortBy") || "")
-  const [sortOrder, setSortOrder] = useState(queryParams.get("sortOrder") || "asc")
 
   const [newPost, setNewPost] = useState({ title: "", body: "", userId: 1 })
 
   const [tags, setTags] = useState([])
-  const [selectedTag, setSelectedTag] = useState(queryParams.get("tag") || "")
 
   const { posts, loading } = usePosts(selectedTag, skip, limit, sortBy, sortOrder)
 
@@ -46,19 +39,10 @@ const PostsManager = () => {
   const [showEditCommentDialog, setShowEditCommentDialog] = useState(false)
   const [showPostDetailDialog, setShowPostDetailDialog] = useState(false)
   const [showUserModal, setShowUserModal] = useState(false)
+
   const [selectedUser, setSelectedUser] = useState(null)
 
   // URL 업데이트 함수
-  const updateURL = () => {
-    const params = new URLSearchParams()
-    if (skip) params.set("skip", skip.toString())
-    if (limit) params.set("limit", limit.toString())
-    if (searchQuery) params.set("search", searchQuery)
-    if (sortBy) params.set("sortBy", sortBy)
-    if (sortOrder) params.set("sortOrder", sortOrder)
-    if (selectedTag) params.set("tag", selectedTag)
-    navigate(`?${params.toString()}`)
-  }
 
   // 태그 가져오기
   const fetchTags = async () => {
@@ -73,20 +57,6 @@ const PostsManager = () => {
   useEffect(() => {
     fetchTags()
   }, [])
-
-  useEffect(() => {
-    updateURL()
-  }, [])
-
-  useEffect(() => {
-    const params = new URLSearchParams(location.search)
-    setSkip(parseInt(params.get("skip") || "0"))
-    setLimit(parseInt(params.get("limit") || "10"))
-    setSearchQuery(params.get("search") || "")
-    setSortBy(params.get("sortBy") || "")
-    setSortOrder(params.get("sortOrder") || "asc")
-    setSelectedTag(params.get("tag") || "")
-  }, [location.search])
 
   return (
     <Card className="w-full max-w-6xl mx-auto">
